+++ generators/src/org/jetbrains/kotlin/generators/tests/GenerateTests.kt
@@ -773,6 +773,7 @@ fun main(args: Array<String>) {
             model("refactoring/introduceLambdaParameter", pattern = KT_OR_KTS, testMethod = "doIntroduceLambdaParameterTest")
             model("refactoring/introduceJavaParameter", extension = "java", testMethod = "doIntroduceJavaParameterTest")
             model("refactoring/introduceTypeParameter", pattern = KT_OR_KTS, testMethod = "doIntroduceTypeParameterTest")
+            model("refactoring/extractSuperclass", pattern = KT_OR_KTS, testMethod = "doExtractSuperclassTest")
         }
 
         testClass<AbstractPullUpTest>() {