+++ idea/src/org/jetbrains/kotlin/idea/actions/NewKotlinFileAction.kt
@@ -82,58 +82,63 @@ class NewKotlinFileAction
         return obj is NewKotlinFileAction
     }
 
-    override fun createFileFromTemplate(name: String, template: FileTemplate, dir: PsiDirectory): PsiFile? {
-        val directorySeparators = if (template.name == "Kotlin File") arrayOf('/', '\\') else arrayOf('/', '\\', '.')
-        val (className, targetDir) = findOrCreateTarget(dir, name, directorySeparators)
+    override fun createFileFromTemplate(name: String, template: FileTemplate, dir: PsiDirectory) =
+            Companion.createFileFromTemplate(name, template, dir)
 
-        val service = DumbService.getInstance(dir.project)
-        service.isAlternativeResolveEnabled = true
-        try {
-            return createFromTemplate(targetDir, className, template)
-        }
-        finally {
-            service.isAlternativeResolveEnabled = false
-        }
-    }
+    companion object {
+        private fun findOrCreateTarget(dir: PsiDirectory, name: String, directorySeparators: Array<Char>): Pair<String, PsiDirectory> {
+            var className = name.removeSuffix(".kt")
+            var targetDir = dir
 
-    private fun findOrCreateTarget(dir: PsiDirectory, name: String, directorySeparators: Array<Char>): Pair<String, PsiDirectory> {
-        var className = name.removeSuffix(".kt")
-        var targetDir = dir
+            for (splitChar in directorySeparators) {
+                if (splitChar in className) {
+                    val names = className.trim().split(splitChar)
 
-        for (splitChar in directorySeparators) {
-            if (splitChar in className) {
-                val names = className.trim().split(splitChar)
+                    for (dirName in names.dropLast(1)) {
+                        targetDir = targetDir.findSubdirectory(dirName) ?: targetDir.createSubdirectory(dirName)
+                    }
 
-                for (dirName in names.dropLast(1)) {
-                    targetDir = targetDir.findSubdirectory(dirName) ?: targetDir.createSubdirectory(dirName)
+                    className = names.last()
+                    break
                 }
+            }
+            return Pair(className, targetDir)
+        }
 
-                className = names.last()
-                break
+        private fun createFromTemplate(dir: PsiDirectory, className: String, template: FileTemplate): PsiFile? {
+            val project = dir.project
+            val defaultProperties = FileTemplateManager.getInstance(project).defaultProperties
+
+            val properties = Properties(defaultProperties)
+
+            val element = try {
+                CreateFromTemplateDialog(project, dir, template,
+                                         AttributesDefaults(className).withFixedName(true),
+                                         properties).create()
+            }
+            catch (e: IncorrectOperationException) {
+                throw e
+            }
+            catch (e: Exception) {
+                LOG.error(e)
+                return null
+            }
+
+            return element?.containingFile
+        }
+
+        fun createFileFromTemplate(name: String, template: FileTemplate, dir: PsiDirectory): PsiFile? {
+            val directorySeparators = if (template.name == "Kotlin File") arrayOf('/', '\\') else arrayOf('/', '\\', '.')
+            val (className, targetDir) = findOrCreateTarget(dir, name, directorySeparators)
+
+            val service = DumbService.getInstance(dir.project)
+            service.isAlternativeResolveEnabled = true
+            try {
+                return createFromTemplate(targetDir, className, template)
+            }
+            finally {
+                service.isAlternativeResolveEnabled = false
             }
         }
-        return Pair(className, targetDir)
-    }
-
-    private fun createFromTemplate(dir: PsiDirectory, className: String, template: FileTemplate): PsiFile? {
-        val project = dir.project
-        val defaultProperties = FileTemplateManager.getInstance(project).defaultProperties
-
-        val properties = Properties(defaultProperties)
-
-        val element = try {
-            CreateFromTemplateDialog(project, dir, template,
-                                     AttributesDefaults(className).withFixedName(true),
-                                     properties).create()
-        }
-        catch (e: IncorrectOperationException) {
-            throw e
-        }
-        catch (e: Exception) {
-            LOG.error(e)
-            return null
-        }
-
-        return element?.containingFile
     }
 }