+++ idea/testData/refactoring/extractSuperclass/enum.kt
@@ -1,0 +1,9 @@
+// NAME: B
+// SIBLING:
+enum class <caret>A {
+    X, Y, Z;
+
+    fun foo() {
+
+    }
+}